security buff
Don't parse ip/ port with json... this just looks like a flaw waiting to be exploited
All errors during authentication must return 401 for security reasons
Also moved doError() into util file, and imported that into app.js (it would have failed lmao)
