+++ lib/api.js
@@ -1,4 +1,4 @@
-const { LimitedMap } = require('./palace-util')
+const { doError, LimitedMap } = require('./palace-util')
 const database = require('./database')
 
 const express = require('express')
@@ -6,46 +6,6 @@ const { Mutex } = require('async-mutex')
 const crypto = require('crypto')
 // Thank you so much H4O, i hate when people use unnecessary dependencies (and fix proto pollution in silly ways)
 const JSONB = require('when-json-met-bigint').JSONB({strict: true, protoAction: 'ignore', constructorAction: 'ignore'})
-
-function doError(res, status, message) {
-    if (!status || typeof status != 'number' || status < 400)
-        status = 418 // I'm a teapot
-    if (!message) {
-        switch (status) {
-            case 400:
-                message = "Bad Request"
-                break
-            case 401:
-                message = "Unauthorized"
-                break
-            case 403:
-                message = "Forbidden"
-                break
-            case 404:
-                message = "Not Found"
-                break
-            case 405:
-                message = "Method Not Allowed"
-                break
-            case 429:
-                message = "Too Many Requests"
-                break
-            case 500:
-                message = "Internal Server Error"
-                break
-            case 501:
-                message = "Not Implemented"
-                break
-            case 503:
-                message = "Service Unavailable"
-                break
-            default:
-                message = "I'm a teapot"
-        }
-    } else if (typeof message != 'string')
-        message = JSON.stringify(message)
-    res.status(status).json({error: 1, message})
-}
 
 function validatePublicToken(req, res, token) {
     // TODO: write this in the future to enable outsiders to access "private" endpoints
@@ -65,7 +25,7 @@ function validateToken(req, res, token) {
     }
     // Only valid for 60 seconds, kindly let server know it is stupid
     if (creation < req.requestTime - 60_000) {
-        doError(res, 400, "Token Expired")
+        doError(res, 401, "Token Expired")
         return false
     }
     return true
@@ -241,6 +201,11 @@ api.get('/player/ignoreList/:uuid', async (req, res) => {
  */
 api.post('/token', async (req, res) => {
 
+    // Small note: We can theoretically return a mix of 4XX codes, however this would give away information to potential
+    // intruders that they've gotten past certain checkpoints, so we cannot do that.
+    // Always return "401 Unauthorized" until ip+tokens match up, in which case the response becomes "401 Token Expired"
+    // just so the server doesn't freak out.
+
     // Ensure sufficient exchange token is set
     if (!process.env.API_TOKEN_SECRET || typeof process.env.API_TOKEN_SECRET != 'string' || process.env.API_TOKEN_SECRET.length != 128)
         return doError(res, 401)
@@ -290,13 +255,13 @@ api.post('/token', async (req, res) => {
                 // clear whole entry, requires exchange to proceed
                 ipTokens[req.remoteExpanded] = null
                 console.error(`Warning! Unexpected REFRESH request from [ ${req.remoteExpanded} ] !!`)
-                return doError(res, 429)
+                return doError(res, 401)
             }
 
             // Ensure we haven't made this request too late, but nicely
             if (req.requestTime >= creation + 60_000) {
                 console.error(`Warning! Late REFRESH request from [ ${req.remoteExpanded} ] !!`)
-                return doError(res, 400, "Token Expired")
+                return doError(res, 401, "Token Expired")
             }
 
         } else {
