+++ lib/palace-util.js
@@ -1,5 +1,52 @@
 
 /**
+ * Automated error responses for express
+ *
+ * @param {Response} res - Express response object
+ * @param {number} status - status code
+ * @param {string} [message] - message response, defaults to related status code
+ */
+function doError(res, status, message) {
+    if (!status || typeof status != 'number' || status < 400)
+        status = 418 // I'm a teapot
+    if (!message) {
+        switch (status) {
+            case 400:
+                message = "Bad Request"
+                break
+            case 401:
+                message = "Unauthorized"
+                break
+            case 403:
+                message = "Forbidden"
+                break
+            case 404:
+                message = "Not Found"
+                break
+            case 405:
+                message = "Method Not Allowed"
+                break
+            case 429:
+                message = "Too Many Requests"
+                break
+            case 500:
+                message = "Internal Server Error"
+                break
+            case 501:
+                message = "Not Implemented"
+                break
+            case 503:
+                message = "Service Unavailable"
+                break
+            default:
+                message = "I'm a teapot"
+        }
+    } else if (typeof message != 'string')
+        message = JSON.stringify(message)
+    res.status(status).json({error: 1, message})
+}
+
+/**
  * Recursively flattens the object and returns a new object.
  *
  * @example <caption>Returns `{"in.depth": 50}`</caption>
@@ -131,6 +178,7 @@ async function sleep(milliseconds) {
 }
 
 module.exports = {
+    doError,
     flattenObject,
     LimitedMap,
     resolveObjectPath,
